+++ dist_k_mean/config.py
@@ -19,7 +19,7 @@ GAUSSIANS_STD_DEV = float(os.getenv('GAUSSIANS_STD_DEV', default=0.1))
 ###### BLACK_BOXES ######
 INNER_BLACKBOX = os.getenv('INNER_BLACKBOX', default='KMeans')  # 'KMeans' 'MiniBatchKMeans' 'ScalableKMeans'
 INNER_BLACKBOX_ITERATIONS = int(os.getenv('INNER_BLACKBOX_ITERATIONS', default=4))
-INNER_BLACKBOX_L_TO_K_RATIO = int(os.getenv('INNER_BLACKBOX_L_TO_K_RATIO', default=1))
+INNER_BLACKBOX_L_TO_K_RATIO = float(os.getenv('INNER_BLACKBOX_L_TO_K_RATIO', default=1))
 
 FINALIZATION_BLACKBOX = os.getenv('FINALIZATION_BLACKBOX', default='KMeans')
